"Modularize traits by splitting into individual files
Refactored `traits.rs` into separate modules: `motor.rs`, `hub.rs`, `thorlabs_device.rs`, and `channel_enable_state.rs`. Improves code maintainability and structure."
